Path change related to Atoms, hamburger renaming
